requirements.txt
Appium-Python-Client==3.2.1
pytest==8.3.4
